Switch to -rdynamic when building on MacOS
bcftools needs to export all of its symbols so that plugins can use them.
There are two ways to do this on MacOS - pass -export_dynamic to the
linker, or -rdynamic to the compiler.  Unfortunately -export_dynamic
isn't supported by older linkers on the Mac, causing build failures.

Using -rdynamic works a bit better.  gcc is fine, but on old systems
clang issues an "argument unused during compilation: '-rdynamic'"
warning.  On newer systems, clang accepts the -rdynamic option without
complaining.  As the warning is non-fatal (as long as -Werror is not
in use) the build completes.

These options only appear to be necessary with clang if link-time
optimization is turned on.  Enabling -flto without -rdynamic or
-Wl,-export_dynamic results in undefined symbols on new MacOS systems.